Methods component is created
